requirements.txt
Django==4.1.3
djangorestframework==3.14.0
drf-yasg==1.21.5
django-filter==23.2
python-dotenv==1.0.0